@@ -13,70 +13,82 @@
   "license": "ISC",
   "dependencies": {
     "@apollo/react-hooks": "^3.1.3",
-    "@babel/plugin-transform-typescript": "^7.6.3",
-    "@babel/preset-typescript": "^7.6.0",
-    "@types/mongodb": "^3.3.6",
-    "@types/mongoose": "^5.5.23",
-    "@types/react": "^16.9.9",
-    "@types/react-dom": "^16.9.3",
+    "@babel/plugin-transform-typescript": "^7.7.4",
+    "@babel/preset-typescript": "^7.7.7",
+    "@types/mongodb": "^3.3.14",
+    "@types/mongoose": "^5.5.39",
+    "@types/react": "^16.9.17",
+    "@types/react-bootstrap": "^0.32.20",
+    "@types/react-dom": "^16.9.4",
     "@types/react-redux": "^7.1.5",
+    "@types/react-router": "^5.1.4",
+    "@types/react-router-dom": "^5.1.3",
     "apollo-boost": "^0.4.7",
+    "bootstrap": "^4.4.1",
     "connect-mongodb-session": "^2.2.0",
-    "copy-webpack-plugin": "^5.0.4",
-    "csstype": "^2.6.7",
-    "eslint-config-prettier": "^6.4.0",
-    "eslint-plugin-prettier": "^3.1.1",
+    "copy-webpack-plugin": "^5.1.1",
+    "csstype": "^2.6.8",
+    "eslint-config-airbnb": "^18.0.1",
+    "eslint-config-prettier": "^6.9.0",
+    "eslint-import-resolver-typescript": "^2.0.0",
+    "eslint-plugin-import": "^2.19.1",
+    "eslint-plugin-json": "^2.0.1",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
     "express": "^4.17.1",
     "express-session": "^1.17.0",
-    "file-loader": "^4.2.0",
+    "file-loader": "^4.3.0",
     "graphql": "^14.5.8",
     "history": "^4.10.1",
+    "jquery": "^3.4.1",
     "less": "^3.10.3",
-    "mongodb": "^3.3.3",
-    "mongoose": "^5.7.7",
+    "mongodb": "^3.4.1",
+    "mongoose": "^5.8.5",
     "passport": "^0.4.1",
     "passport-google-oauth20": "^2.0.0",
-    "prettier-eslint": "^9.0.0",
+    "prettier": "^1.19.1",
+    "prettier-eslint": "^9.0.1",
     "react-apollo": "^3.1.3",
+    "react-bootstrap": "^1.0.0-beta.16",
     "react-css-modules": "^4.7.11",
     "react-google-login": "^5.0.7",
-    "react-redux": "^7.1.1",
+    "react-redux": "^7.1.3",
     "react-router-dom": "^5.1.2",
-    "redux": "^4.0.4",
-    "redux-react-session": "^2.6.0",
+    "redux": "^4.0.5",
+    "redux-react-session": "^2.6.1",
     "reset-css": "^4.0.1",
     "svg-inline-loader": "^0.8.0",
     "svg-loader": "0.0.2",
-    "ts-loader": "^6.2.0",
-    "tslint": "^5.20.0",
+    "ts-loader": "^6.2.1",
+    "tslint": "^5.20.1",
     "tslint-react": "^4.1.0",
-    "typescript": "^3.6.4",
-    "underscore": "^1.9.1",
-    "url-loader": "^2.1.0",
-    "webpack-dev-middleware": "^3.7.1"
+    "typescript": "^3.7.4",
+    "underscore": "^1.9.2",
+    "url-loader": "^2.3.0",
+    "webpack-dev-middleware": "^3.7.2"
   },
   "devDependencies": {
-    "@babel/cli": "^7.6.0",
-    "@babel/core": "^7.6.0",
-    "@babel/node": "^7.6.1",
-    "@babel/preset-env": "^7.6.0",
-    "@babel/preset-react": "^7.0.0",
-    "@typescript-eslint/eslint-plugin": "^2.4.0",
-    "@typescript-eslint/parser": "^2.4.0",
+    "@babel/cli": "^7.7.7",
+    "@babel/core": "^7.7.7",
+    "@babel/node": "^7.7.7",
+    "@babel/preset-env": "^7.7.7",
+    "@babel/preset-react": "^7.7.4",
+    "@typescript-eslint/eslint-plugin": "^2.15.0",
+    "@typescript-eslint/parser": "^2.15.0",
     "babel-loader": "^8.0.6",
     "css": "^2.2.4",
-    "css-loader": "^3.2.0",
-    "eslint": "^6.4.0",
+    "css-loader": "^3.4.1",
+    "eslint": "^6.8.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-prettier": "^3.1.2",
+    "eslint-plugin-react": "^7.17.0",
     "less-loader": "^5.0.0",
-    "mini-css-extract-plugin": "^0.8.0",
-    "nodemon": "^1.19.2",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "style-loader": "^1.0.0",
-    "webpack": "^4.40.2",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.1"
+    "mini-css-extract-plugin": "^0.8.2",
+    "nodemon": "^1.19.4",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "style-loader": "^1.1.2",
+    "webpack": "^4.41.5",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.10.1"
   }
 }
