@@ -2,21 +2,31 @@ import * as React from "react";
 
 import {
   setMetronomeSound,
-  ExclusiveActionTypes
+  ExclusiveActionTypes,
+  SetMetronomeSoundAction
 } from "../../../actions/workspace/workspace";
 
 import { StateType } from "../../../reducers/index";
-import { SettingSelect } from "../../input/setting-select";
+import { SettingSelect } from "../../element/setting-select";
 
-const { connect } = require("react-redux");
+import { connect } from "react-redux";
+import { Dispatch } from "redux";
 
-const mapStateToProps = (state: StateType) => {
+interface StateProps {
+  metronomeSound: string;
+}
+
+interface DispatchProps {
+  setMetronomeSound: (metronomeSound: string) => SetMetronomeSoundAction;
+}
+
+const mapStateToProps = (state: StateType): StateProps => {
   return { metronomeSound: state.workspace.exclusive.metronomeSound };
 };
 
-const mapDispatchToProps = (dispatch: any) => {
+const mapDispatchToProps = (dispatch: Dispatch): DispatchProps => {
   return {
-    setMetronomeSound: (metronomeSound: string) =>
+    setMetronomeSound: (metronomeSound: string): SetMetronomeSoundAction =>
       dispatch(setMetronomeSound(metronomeSound))
   };
 };
@@ -27,11 +37,14 @@ const ConnectedMetronomeSetting = ({
 }: {
   metronomeSound: string;
   setMetronomeSound: (payload: string) => ExclusiveActionTypes;
-}) => {
+}): JSX.Element => {
   return (
     <SettingSelect
       value={metronomeSound}
-      options={[{ index: 1, value: "Hi-Hat" }, { index: 2, value: "Beep" }]}
+      options={[
+        { index: 1, value: "Hi-Hat" },
+        { index: 2, value: "Beep" }
+      ]}
       reduxDispatch={setMetronomeSound}
     ></SettingSelect>
   );
