@@ -7,9 +7,10 @@ import { useQuery } from "@apollo/react-hooks";
 const styles = require("./header.less");
 import { VampLogo } from "./logo";
 
-import { ButtonLinkDefault } from "../input/button";
+import { ButtonLinkDefault } from "../element/button";
+import LoggedInUserButton from "./logged-in-user-button";
 
-const meQuery = gql`
+const ME_QUERY = gql`
   {
     me {
       username
@@ -22,32 +23,29 @@ interface MeUser {
 }
 
 interface MeData {
-  me: MeUser
+  me: MeUser;
 }
 
-interface MeVars {
-}
+interface MeVars {}
 
 const VampHeader = () => {
   // RE APOLLO RE GRAPHQL
   // https://www.apollographql.com/docs/react/development-testing/static-typing/
-  // This is the functional way to do GraphQL queries with TS. Note the 
+  // This is the functional way to do GraphQL queries with TS. Note the
   // interfaces above.
-  const {loading, data} = useQuery<MeData, MeVars>(meQuery, {variables: {}});
+  const { loading, data } = useQuery<MeData, MeVars>(ME_QUERY, {
+    variables: {}
+  });
 
   const button =
     loading || data.me == null ? (
       <ButtonLinkDefault
         text="Log In"
-        style={{ marginTop: "auto", marginBottom: "auto" }}
+        // style={{ marginTop: "auto", marginBottom: "auto" }}
         href="/login"
       ></ButtonLinkDefault>
     ) : (
-      <ButtonLinkDefault
-        text={data.me.username}
-        style={{ marginTop: "auto", marginBottom: "auto" }}
-        href="/login"
-      ></ButtonLinkDefault>
+      <LoggedInUserButton username={data.me.username}></LoggedInUserButton>
     );
 
   return (
@@ -60,4 +58,4 @@ const VampHeader = () => {
   );
 };
 
-export default graphql(meQuery)(VampHeader);
+export default graphql(ME_QUERY)(VampHeader);
